Move C++ Module/Engine types to their own headers
Add some #ifdef guards for Wasmtime features as well as some small tests.
