+++ crates/c-api/include/wasmtime.hh
@@ -41,154 +41,15 @@
 #include <ostream>
 #include <variant>
 #include <vector>
-#ifdef __has_include
-#if __has_include(<span>)
-#include <span>
-#endif
-#endif
 
 #include <wasmtime.h>
 #include <wasmtime/config.hh>
+#include <wasmtime/engine.hh>
 #include <wasmtime/error.hh>
+#include <wasmtime/module.hh>
 #include <wasmtime/types.hh>
 
 namespace wasmtime {
-
-#ifdef __cpp_lib_span
-
-/// \brief Alias to C++20 std::span when it is available
-template <typename T, std::size_t Extent = std::dynamic_extent>
-using Span = std::span<T, Extent>;
-
-#else
-
-/// \brief Means number of elements determined at runtime
-inline constexpr size_t dynamic_extent =
-    std::numeric_limits<std::size_t>::max();
-
-/**
- * \brief Span class used when c++20 is not available
- * @tparam T Type of data
- * @tparam Extent Static size of data referred by Span class
- */
-template <typename T, std::size_t Extent = dynamic_extent> class Span;
-
-/// \brief Check whether a type is `Span`
-template <typename T> struct IsSpan : std::false_type {};
-
-template <typename T, std::size_t Extent>
-struct IsSpan<Span<T, Extent>> : std::true_type {};
-
-template <typename T, std::size_t Extent> class Span {
-  static_assert(Extent == dynamic_extent,
-                "The current implementation supports dynamic-extent span only");
-
-public:
-  /// \brief Type used to iterate over this span (a raw pointer)
-  using iterator = T *;
-
-  /// \brief Constructor of Span class
-  Span(T *t, std::size_t n) : ptr_{t}, size_{n} {}
-
-  /// \brief Constructor of Span class for containers
-  template <typename C,
-            std::enable_if_t<
-                !IsSpan<C>::value &&
-                    std::is_pointer_v<decltype(std::declval<C &>().data())> &&
-                    std::is_convertible_v<
-                        std::remove_pointer_t<
-                            decltype(std::declval<C &>().data())> (*)[],
-                        T (*)[]> &&
-                    std::is_convertible_v<decltype(std::declval<C>().size()),
-                                          std::size_t>,
-                int> = 0>
-  Span(C &range) : ptr_{range.data()}, size_{range.size()} {}
-
-  /// \brief Returns item by index
-  T &operator[](ptrdiff_t idx) const {
-    return ptr_[idx]; // NOLINT
-  }
-
-  /// \brief Returns pointer to data
-  T *data() const { return ptr_; }
-
-  /// \brief Returns number of data that referred by Span class
-  std::size_t size() const { return size_; }
-
-  /// \brief Returns begin iterator
-  iterator begin() const { return ptr_; }
-
-  /// \brief Returns end iterator
-  iterator end() const {
-    return ptr_ + size_; // NOLINT
-  }
-
-  /// \brief Returns size in bytes
-  std::size_t size_bytes() const { return sizeof(T) * size_; }
-
-private:
-  T *ptr_;
-  std::size_t size_;
-};
-
-#endif
-
-
-/**
- * \brief Global compilation state in Wasmtime.
- *
- * Created with either default configuration or with a specified instance of
- * configuration, an `Engine` is used as an umbrella "session" for all other
- * operations in Wasmtime.
- */
-class Engine {
-  friend class Store;
-  friend class Module;
-  friend class Linker;
-
-  struct deleter {
-    void operator()(wasm_engine_t *p) const { wasm_engine_delete(p); }
-  };
-
-  std::unique_ptr<wasm_engine_t, deleter> ptr;
-
-public:
-  /// \brief Creates an engine with default compilation settings.
-  Engine() : ptr(wasm_engine_new()) {}
-  /// \brief Creates an engine with the specified compilation settings.
-  explicit Engine(Config config)
-      : ptr(wasm_engine_new_with_config(config.ptr.release())) {}
-
-  /// \brief Increments the current epoch which may result in interrupting
-  /// currently executing WebAssembly in connected stores if the epoch is now
-  /// beyond the configured threshold.
-  void increment_epoch() const { wasmtime_engine_increment_epoch(ptr.get()); }
-};
-
-/**
- * \brief Converts the WebAssembly text format into the WebAssembly binary
- * format.
- *
- * This will parse the text format and attempt to translate it to the binary
- * format. Note that the text parser assumes that all WebAssembly features are
- * enabled and will parse syntax of future proposals. The exact syntax here
- * parsed may be tweaked over time.
- *
- * Returns either an error if parsing failed or the wasm binary.
- */
-inline Result<std::vector<uint8_t>> wat2wasm(std::string_view wat) {
-  wasm_byte_vec_t ret;
-  auto *error = wasmtime_wat2wasm(wat.data(), wat.size(), &ret);
-  if (error != nullptr) {
-    return Error(error);
-  }
-  std::vector<uint8_t> vec;
-  // NOLINTNEXTLINE TODO can this be done without triggering lints?
-  Span<uint8_t> raw(reinterpret_cast<uint8_t *>(ret.data), ret.size);
-  vec.assign(raw.begin(), raw.end());
-  wasm_byte_vec_delete(&ret);
-  return vec;
-}
 
 /**
  * \brief Non-owning reference to a WebAssembly function frame as part of a
@@ -353,171 +214,6 @@ struct TrapError {
 /// Result used by functions which can fail because of invariants being violated
 /// (such as a type error) as well as because of a WebAssembly trap.
 template <typename T> using TrapResult = Result<T, TrapError>;
-
-/**
- * \brief Representation of a compiled WebAssembly module.
- *
- * This type contains JIT code of a compiled WebAssembly module. A `Module` is
- * connected to an `Engine` and can only be instantiated within that `Engine`.
- * You can inspect a `Module` for its type information. This is passed as an
- * argument to other APIs to instantiate it.
- */
-class Module {
-  friend class Store;
-  friend class Instance;
-  friend class Linker;
-
-  struct deleter {
-    void operator()(wasmtime_module_t *p) const { wasmtime_module_delete(p); }
-  };
-
-  std::unique_ptr<wasmtime_module_t, deleter> ptr;
-
-  Module(wasmtime_module_t *raw) : ptr(raw) {}
-
-public:
-  /// Copies another module into this one.
-  Module(const Module &other) : ptr(wasmtime_module_clone(other.ptr.get())) {}
-  /// Copies another module into this one.
-  Module &operator=(const Module &other) {
-    ptr.reset(wasmtime_module_clone(other.ptr.get()));
-    return *this;
-  }
-  ~Module() = default;
-  /// Moves resources from another module into this one.
-  Module(Module &&other) = default;
-  /// Moves resources from another module into this one.
-  Module &operator=(Module &&other) = default;
-
-  /**
-   * \brief Compiles a module from the WebAssembly text format.
-   *
-   * This function will automatically use `wat2wasm` on the input and then
-   * delegate to the #compile function.
-   */
-  static Result<Module> compile(Engine &engine, std::string_view wat) {
-    auto wasm = wat2wasm(wat);
-    if (!wasm) {
-      return wasm.err();
-    }
-    auto bytes = wasm.ok();
-    return compile(engine, bytes);
-  }
-
-  /**
-   * \brief Compiles a module from the WebAssembly binary format.
-   *
-   * This function compiles the provided WebAssembly binary specified by `wasm`
-   * within the compilation settings configured by `engine`. This method is
-   * synchronous and will not return until the module has finished compiling.
-   *
-   * This function can fail if the WebAssembly binary is invalid or doesn't
-   * validate (or similar).
-   */
-  static Result<Module> compile(Engine &engine, Span<uint8_t> wasm) {
-    wasmtime_module_t *ret = nullptr;
-    auto *error =
-        wasmtime_module_new(engine.ptr.get(), wasm.data(), wasm.size(), &ret);
-    if (error != nullptr) {
-      return Error(error);
-    }
-    return Module(ret);
-  }
-
-  /**
-   * \brief Validates the provided WebAssembly binary without compiling it.
-   *
-   * This function will validate whether the provided binary is indeed valid
-   * within the compilation settings of the `engine` provided.
-   */
-  static Result<std::monostate> validate(Engine &engine, Span<uint8_t> wasm) {
-    auto *error =
-        wasmtime_module_validate(engine.ptr.get(), wasm.data(), wasm.size());
-    if (error != nullptr) {
-      return Error(error);
-    }
-    return std::monostate();
-  }
-
-  /**
-   * \brief Deserializes a previous list of bytes created with `serialize`.
-   *
-   * This function is intended to be much faster than `compile` where it uses
-   * the artifacts of a previous compilation to quickly create an in-memory
-   * module ready for instantiation.
-   *
-   * It is not safe to pass arbitrary input to this function, it is only safe to
-   * pass in output from previous calls to `serialize`. For more information see
-   * the Rust documentation -
-   * https://docs.wasmtime.dev/api/wasmtime/struct.Module.html#method.deserialize
-   */
-  static Result<Module> deserialize(Engine &engine, Span<uint8_t> wasm) {
-    wasmtime_module_t *ret = nullptr;
-    auto *error = wasmtime_module_deserialize(engine.ptr.get(), wasm.data(),
-                                              wasm.size(), &ret);
-    if (error != nullptr) {
-      return Error(error);
-    }
-    return Module(ret);
-  }
-
-  /**
-   * \brief Deserializes a module from an on-disk file.
-   *
-   * This function is the same as `deserialize` except that it reads the data
-   * for the serialized module from the path on disk. This can be faster than
-   * the alternative which may require copying the data around.
-   *
-   * It is not safe to pass arbitrary input to this function, it is only safe to
-   * pass in output from previous calls to `serialize`. For more information see
-   * the Rust documentation -
-   * https://docs.wasmtime.dev/api/wasmtime/struct.Module.html#method.deserialize
-   */
-  static Result<Module> deserialize_file(Engine &engine,
-                                         const std::string &path) {
-    wasmtime_module_t *ret = nullptr;
-    auto *error =
-        wasmtime_module_deserialize_file(engine.ptr.get(), path.c_str(), &ret);
-    if (error != nullptr) {
-      return Error(error);
-    }
-    return Module(ret);
-  }
-
-  /// Returns the list of types imported by this module.
-  ImportType::List imports() const {
-    ImportType::List list;
-    wasmtime_module_imports(ptr.get(), &list.list);
-    return list;
-  }
-
-  /// Returns the list of types exported by this module.
-  ExportType::List exports() const {
-    ExportType::List list;
-    wasmtime_module_exports(ptr.get(), &list.list);
-    return list;
-  }
-
-  /**
-   * \brief Serializes this module to a list of bytes.
-   *
-   * The returned bytes can then be used to later pass to `deserialize` to
-   * quickly recreate this module in a different process perhaps.
-   */
-  Result<std::vector<uint8_t>> serialize() const {
-    wasm_byte_vec_t bytes;
-    auto *error = wasmtime_module_serialize(ptr.get(), &bytes);
-    if (error != nullptr) {
-      return Error(error);
-    }
-    std::vector<uint8_t> ret;
-    // NOLINTNEXTLINE TODO can this be done without triggering lints?
-    Span<uint8_t> raw(reinterpret_cast<uint8_t *>(bytes.data), bytes.size);
-    ret.assign(raw.begin(), raw.end());
-    wasm_byte_vec_delete(&bytes);
-    return ret;
-  }
-};
 
 /**
  * \brief Configuration for an instance of WASI.
